refactor: split four large source files into focused modules
The clud build warning flagged four files >=40 kB that AI tooling
struggles to work with. This is a pure MOVE refactor (no behavior
changes) splitting each into smaller sibling modules. Public APIs
preserved via re-exports / module wiring; all callers untouched.

crates/fastled-cli/src/lib.rs (52 kB -> 13 kB)
  + cli.rs, selection.rs, compile_stream.rs, commands.rs,
    dwarf_smoke.rs, paths_util.rs
  + entries added to dylints/ban_std_pathbuf/src/allowlist.txt for the
    new sibling modules (they inherit raw PathBuf usage from lib.rs)

src/fastled/frontend/index.ts (57 kB -> 4.5 kB)
  + bootstrap.ts, logging_setup.ts, vfs_bridge.ts, canvas_render.ts,
    fastled_init.ts, video_recording_init.ts, debug_api.ts, state.ts

src/fastled/frontend/modules/audio/audio_manager.ts (62 kB -> 40 kB)
  + audio_constants.ts, audio_processor.ts, audio_debug_api.ts

src/fastled/frontend/modules/ui/ui_manager.ts (80 kB -> 58 kB)
  + ui_control_factory.ts, ui_help_renderer.ts, ui_debug_api.ts

Verified: soldr cargo test -p fastled-cli (129 passed + 1 doctest),
bash test (13 pytest passed), tsc --noEmit clean.

Known: ui_manager.ts remains 58 kB (down from 80) -- the residual
is the JsonUiManager class itself; deeper class-method extraction
deferred to a follow-up.
