+++ src/fastled/frontend/modules/ui/ui_manager.ts
@@ -44,6 +44,19 @@
 import { AudioManager } from '../audio/audio_manager.ts';
 import { UILayoutPlacementManager } from './ui_layout_placement_manager.ts';
 import { UIRecorder } from '../recording/ui_recorder.ts';
+import {
+  createButton,
+  createCheckbox,
+  createDropdown,
+  createNumberField,
+  createNumberFieldPair,
+  createSlider,
+  groupAdjacentNumberFields,
+  setDescription,
+  setTitle,
+} from './ui_control_factory.ts';
+import { createHelp } from './ui_help_renderer.ts';
+import { installUiDebugApi } from './ui_debug_api.ts';
 
 /** Global instance of AudioManager for audio processing */
 const audioManager = new AudioManager();
@@ -54,745 +67,14 @@ window.setupAudioAnalysis = function (audioElement) {
 };
 
 /**
- * Simple markdown to HTML converter
- * Supports: headers, bold, italic, code, links, lists, and paragraphs
- * @param {string} markdown - Markdown text to convert
- * @returns {string} HTML string with converted markdown
- */
-function markdownToHtml(markdown) {
-  if (!markdown) return '';
-
-  let html = markdown;
-
-  // Convert headers (# ## ### etc.)
-  html = html.replace(/^### (.+)$/gm, '<h3>$1</h3>');
-  html = html.replace(/^## (.+)$/gm, '<h2>$1</h2>');
-  html = html.replace(/^# (.+)$/gm, '<h1>$1</h1>');
-
-  // Convert bold **text** and __text__
-  html = html.replace(/\*\*(.+?)\*\*/g, '<strong>$1</strong>');
-  html = html.replace(/__(.+?)__/g, '<strong>$1</strong>');
-
-  // Convert italic *text* and _text_
-  html = html.replace(/\*(.+?)\*/g, '<em>$1</em>');
-  html = html.replace(/_(.+?)_/g, '<em>$1</em>');
-
-  // Convert inline code `code`
-  html = html.replace(/`([^`]+)`/g, '<code>$1</code>');
-
-  // Convert code blocks ```code```
-  html = html.replace(/```([^`]+)```/g, '<pre><code>$1</code></pre>');
-
-  // Convert links [text](url)
-  html = html.replace(/\[([^\]]+)\]\(([^)]+)\)/g, '<a href="$2" target="_blank">$1</a>');
-
-  // Convert unordered lists (- item or * item)
-  html = html.replace(/^[\-\*] (.+)$/gm, '<li>$1</li>');
-
-  // Convert ordered lists (1. item, 2. item, etc.)
-  html = html.replace(/^\d+\. (.+)$/gm, '<li class="ordered">$1</li>');
-
-  // Wrap consecutive <li> elements properly
-  html = html.replace(
-    /(<li(?:\s+class="ordered")?>.*?<\/li>(?:\s*<li(?:\s+class="ordered")?>.*?<\/li>)*)/gs,
-    (match) => {
-      if (match.includes('class="ordered"')) {
-        return `<ol>${match.replace(/\s+class="ordered"/g, '')}</ol>`;
-      }
-      return `<ul>${match}</ul>`;
-    },
-  );
-
-  // Convert line breaks to paragraphs (double newlines become paragraph breaks)
-  const paragraphs = html.split(/\n\s*\n/);
-  html = paragraphs.map((p) => {
-    const trimmed = p.trim();
-    if (
-      trimmed && !trimmed.startsWith('<h') && !trimmed.startsWith('<ul')
-      && !trimmed.startsWith('<ol') && !trimmed.startsWith('<pre')
-    ) {
-      return `<p>${trimmed.replace(/\n/g, '<br>')}</p>`;
-    }
-    return trimmed;
-  }).join('\n');
-
-  return html;
-}
-
-/**
- * Groups adjacent number fields into pairs for more efficient space utilization
- * @param {Array} elements - Array of UI element configurations
- * @returns {Array} Array with adjacent number fields grouped into pairs
- */
-function groupAdjacentNumberFields(elements) {
-  const result = [];
-  let i = 0;
-
-  while (i < elements.length) {
-    const current = elements[i];
-
-    // Check if current element is a number field and the next one is also a number field
-    if (current.type === 'number' && i + 1 < elements.length && elements[i + 1].type === 'number') {
-      const next = elements[i + 1];
-
-      // Create a paired element that will be handled specially
-      result.push({
-        type: 'number-pair',
-        leftElement: current,
-        rightElement: next,
-        id: `pair-${current.id}-${next.id}`,
-        group: current.group, // Use the group from the first element
-      });
-
-      i += 2; // Skip both elements since we've paired them
-    } else {
-      // Add single element as-is
-      result.push(current);
-      i += 1;
-    }
-  }
-
-  return result;
-}
-
-/**
- * Creates a number input field UI element
- * @param {Object} element - Element configuration object
- * @param {string} element.name - Display name for the input
- * @param {string} element.id - Unique identifier for the input
- * @param {number} element.value - Current/default value
- * @param {number} element.min - Minimum allowed value
- * @param {number} element.max - Maximum allowed value
- * @param {number} [element.step] - Step increment (defaults to 'any')
- * @returns {HTMLDivElement} Container div with label and number input
- */
-function createNumberField(element) {
-  return createSingleNumberField(element);
-}
-
-/**
- * Creates a paired number input field UI element with two number fields side by side
- * @param {Object} leftElement - Left element configuration object
- * @param {Object} rightElement - Right element configuration object
- * @returns {HTMLDivElement} Container div with two number fields in a flex layout
- */
-function createNumberFieldPair(leftElement, rightElement) {
-  const pairContainer = document.createElement('div');
-  pairContainer.className = 'ui-control number-pair-control';
-  pairContainer.style.display = 'flex';
-  pairContainer.style.gap = '20px';
-  pairContainer.style.alignItems = 'center';
-  pairContainer.style.justifyContent = 'space-between';
-
-  // Create left number field
-  const leftField = createSingleNumberField(leftElement);
-  leftField.style.flex = '1';
-  leftField.style.maxWidth = 'calc(50% - 10px)';
-
-  // Create right number field
-  const rightField = createSingleNumberField(rightElement);
-  rightField.style.flex = '1';
-  rightField.style.maxWidth = 'calc(50% - 10px)';
-
-  pairContainer.appendChild(leftField);
-  pairContainer.appendChild(rightField);
-
-  // Store reference to both elements for later registration
-  /** @type {HTMLDivElement & {_leftElement?: any, _rightElement?: any, _leftControl?: any, _rightControl?: any}} */
-  const containerWithProps = /** @type {any} */ (pairContainer);
-  containerWithProps._leftElement = leftElement;
-  containerWithProps._rightElement = rightElement;
-  containerWithProps._leftControl = leftField;
-  containerWithProps._rightControl = rightField;
-
-  return pairContainer;
-}
-
-/**
- * Creates a single number input field UI element (used by both single and paired fields)
- * @param {Object} element - Element configuration object
- * @returns {HTMLDivElement} Container div with label and number input
- */
-function createSingleNumberField(element) {
-  const controlDiv = document.createElement('div');
-  controlDiv.className = 'ui-control number-control inline-row single-number-field';
-
-  const label = document.createElement('label');
-  label.textContent = element.name;
-  label.htmlFor = `number-${element.id}`;
-  label.style.display = 'inline-block';
-  label.style.verticalAlign = 'middle';
-  label.style.fontWeight = '500';
-  label.style.color = '#E0E0E0';
-  label.style.marginRight = '10px';
-  label.style.fontSize = '0.9em'; // Slightly smaller for paired layout
-
-  const numberInput = document.createElement('input');
-  numberInput.type = 'number';
-  numberInput.id = `number-${element.id}`;
-  numberInput.value = element.value;
-  numberInput.min = element.min;
-  numberInput.max = element.max;
-  numberInput.step = (element.step !== undefined) ? element.step : 'any';
-  numberInput.style.display = 'inline-block';
-  numberInput.style.verticalAlign = 'middle';
-  numberInput.style.boxSizing = 'border-box';
-
-  // Calculate dynamic width based on expected value range
-  const calculateInputWidth = () => {
-    const minDigits = String(Math.floor(element.min || 0)).length;
-    const maxDigits = String(Math.floor(element.max || 0)).length;
-    const valueDigits = String(Math.floor(element.value || 0)).length;
-    const maxLength = Math.max(minDigits, maxDigits, valueDigits, 5); // Minimum 5 digits
-    // ~8px per character + 20px for padding/controls
-    return `${maxLength * 8 + 20}px`;
-  };
-
-  numberInput.style.width = calculateInputWidth();
-  numberInput.style.minWidth = '70px'; // Ensure minimum usable width
-
-  controlDiv.appendChild(label);
-  controlDiv.appendChild(numberInput);
-
-  return controlDiv;
-}
-
-/**
- * Creates an audio input field UI element
- * Delegates to the AudioManager for specialized audio handling
+ * Internal helper preserved at the original call site in `createControlElement`.
+ * Delegates to the module-level `audioManager` singleton — kept here so we don't
+ * have to share that instance across modules.
  * @param {Object} element - Element configuration object for audio control
  * @returns {HTMLElement} Audio control element from AudioManager
  */
 function createAudioField(element) {
   return audioManager.createAudioField(element);
-}
-
-/**
- * Creates a slider (range) input UI element with live value display
- * @param {Object} element - Element configuration object
- * @param {string} element.name - Display name for the slider
- * @param {string} element.id - Unique identifier for the slider
- * @param {number} element.value - Current/default value
- * @param {number} element.min - Minimum allowed value
- * @param {number} element.max - Maximum allowed value
- * @param {number} element.step - Step increment for the slider
- * @returns {HTMLDivElement} Container div with label, value display, and slider
- */
-function createSlider(element) {
-  const controlDiv = document.createElement('div');
-  controlDiv.className = 'ui-control slider-control';
-
-  // Create the slider container with relative positioning
-  const sliderContainer = document.createElement('div');
-  sliderContainer.className = 'slider-container';
-
-  // Create the slider input
-  const slider = document.createElement('input');
-  slider.type = 'range';
-  slider.id = `slider-${element.id}`;
-  slider.setAttribute('min', String(Number.parseFloat(String(element.min))));
-  slider.setAttribute('max', String(Number.parseFloat(String(element.max))));
-  slider.setAttribute('value', String(Number.parseFloat(String(element.value))));
-
-  // Check if element.step exists and is not undefined/null
-  if (element.step !== undefined && element.step !== null) {
-    slider.setAttribute('step', String(Number.parseFloat(String(element.step))));
-  } else {
-    // Set a default step
-    slider.step = 'any';
-  }
-
-  // Create the overlay label div
-  const overlayDiv = document.createElement('div');
-  overlayDiv.className = 'slider-label-overlay';
-
-  const labelText = document.createElement('span');
-  labelText.className = 'label-text';
-  labelText.textContent = element.name;
-
-  const valueDisplay = document.createElement('span');
-  valueDisplay.className = 'slider-value';
-  valueDisplay.textContent = Number(element.value).toFixed(3);
-
-  overlayDiv.appendChild(labelText);
-  overlayDiv.appendChild(valueDisplay);
-
-  // Set initial value in next frame to ensure proper initialization
-  setTimeout(() => {
-    slider.setAttribute('value', String(Number.parseFloat(String(element.value))));
-    valueDisplay.textContent = Number(slider.value).toFixed(3);
-  }, 0);
-
-  // Update value display when slider changes
-  slider.addEventListener('input', () => {
-    valueDisplay.textContent = Number(slider.value).toFixed(3);
-  });
-
-  // Add elements to container
-  sliderContainer.appendChild(slider);
-  sliderContainer.appendChild(overlayDiv);
-  controlDiv.appendChild(sliderContainer);
-
-  return controlDiv;
-}
-
-/**
- * Creates a checkbox UI element for boolean values
- * @param {Object} element - Element configuration object
- * @param {string} element.name - Display name for the checkbox
- * @param {string} element.id - Unique identifier for the checkbox
- * @param {boolean} element.value - Current/default checked state
- * @returns {HTMLDivElement} Container div with label and checkbox
- */
-function createCheckbox(element) {
-  const controlDiv = document.createElement('div');
-  controlDiv.className = 'ui-control';
-
-  const label = document.createElement('label');
-  label.textContent = element.name;
-  label.htmlFor = `checkbox-${element.id}`;
-
-  const checkbox = document.createElement('input');
-  checkbox.type = 'checkbox';
-  checkbox.id = `checkbox-${element.id}`;
-  checkbox.checked = element.value;
-
-  const flexContainer = document.createElement('div');
-  flexContainer.style.display = 'flex';
-  flexContainer.style.alignItems = 'center';
-  flexContainer.style.justifyContent = 'space-between';
-
-  flexContainer.appendChild(label);
-  flexContainer.appendChild(checkbox);
-
-  controlDiv.appendChild(flexContainer);
-
-  return controlDiv;
-}
-
-/**
- * Creates a button UI element with press/release state tracking
- * @param {Object} element - Element configuration object
- * @param {string} element.name - Display text for the button
- * @param {string} element.id - Unique identifier for the button
- * @returns {HTMLDivElement} Container div with configured button element
- */
-function createButton(element) {
-  const controlDiv = document.createElement('div');
-  controlDiv.className = 'ui-control';
-
-  const button = document.createElement('button');
-  button.textContent = element.name;
-  button.id = `button-${element.id}`;
-  button.setAttribute('data-pressed', 'false');
-
-  button.addEventListener('mousedown', () => {
-    button.setAttribute('data-pressed', 'true');
-    button.classList.add('active');
-  });
-
-  button.addEventListener('mouseup', () => {
-    button.setAttribute('data-pressed', 'false');
-    button.classList.remove('active');
-  });
-
-  button.addEventListener('mouseleave', () => {
-    button.setAttribute('data-pressed', 'false');
-    button.classList.remove('active');
-  });
-  controlDiv.appendChild(button);
-  return controlDiv;
-}
-
-function createDropdown(element) {
-  const controlDiv = document.createElement('div');
-  controlDiv.className = 'ui-control';
-
-  const label = document.createElement('label');
-  label.textContent = element.name;
-  label.htmlFor = `dropdown-${element.id}`;
-
-  const dropdown = document.createElement('select');
-  dropdown.id = `dropdown-${element.id}`;
-  dropdown.value = element.value;
-
-  // Add options to the dropdown
-  if (element.options && Array.isArray(element.options)) {
-    element.options.forEach((option, index) => {
-      const optionElement = document.createElement('option');
-      optionElement.value = index;
-      optionElement.textContent = option;
-      dropdown.appendChild(optionElement);
-    });
-  }
-
-  // Set the selected option
-  dropdown.selectedIndex = element.value;
-
-  controlDiv.appendChild(label);
-  controlDiv.appendChild(dropdown);
-
-  return controlDiv;
-}
-
-function setTitle(titleData) {
-  if (titleData && titleData.text) {
-    document.title = titleData.text;
-    const h1Element = document.querySelector('h1');
-    if (h1Element) {
-      h1Element.textContent = titleData.text;
-    } else {
-      console.warn('H1 element not found in document');
-    }
-  } else {
-    console.warn('Invalid title data received:', titleData);
-  }
-}
-
-function setDescription(descData) {
-  if (descData && descData.text) {
-    // Create or find description element
-    let descElement = document.querySelector('#fastled-description');
-    if (!descElement) {
-      descElement = document.createElement('div');
-      descElement.id = 'fastled-description';
-      // Insert after h1
-      const h1Element = document.querySelector('h1');
-      if (h1Element && h1Element.nextSibling) {
-        h1Element.parentNode.insertBefore(descElement, h1Element.nextSibling);
-      } else {
-        console.warn('Could not find h1 element to insert description after');
-        document.body.insertBefore(descElement, document.body.firstChild);
-      }
-    }
-
-    // Always process text as markdown (plain text is valid markdown)
-    descElement.innerHTML = markdownToHtml(descData.text);
-  } else {
-    console.warn('Invalid description data received:', descData);
-  }
-}
-
-function createHelp(element) {
-  const helpContainer = document.createElement('div');
-  helpContainer.className = 'ui-help-container';
-  helpContainer.id = `help-${element.id}`;
-
-  // Create help button
-  const helpButton = document.createElement('button');
-  helpButton.className = 'ui-help-button';
-  helpButton.textContent = '?';
-  helpButton.setAttribute('type', 'button');
-  helpButton.setAttribute('aria-label', 'Help');
-
-  // Prepare content for tooltip and popup
-  const markdownContent = element.markdownContent || '';
-  const tooltipText = markdownContent.length > 200
-    ? `${markdownContent.substring(0, 200).trim()}...`
-    : markdownContent;
-
-  // Convert markdown to HTML for popup
-  const htmlContent = markdownToHtml(markdownContent);
-
-  // Create tooltip
-  const tooltip = document.createElement('div');
-  tooltip.className = 'ui-help-tooltip';
-  tooltip.textContent = tooltipText;
-
-  // Add event listeners for tooltip
-  helpButton.addEventListener('mouseenter', () => {
-    if (tooltipText.trim()) {
-      showTooltip(helpButton, tooltip);
-    }
-  });
-
-  helpButton.addEventListener('mouseleave', () => {
-    hideTooltip(tooltip);
-  });
-
-  // Add event listener for popup
-  helpButton.addEventListener('click', (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    showHelpPopup(htmlContent);
-  });
-
-  // Assemble the help container
-  helpContainer.appendChild(helpButton);
-
-  // Append tooltip to document body so it can appear above everything
-  document.body.appendChild(tooltip);
-
-  // Add styles if not already present
-  if (!document.querySelector('#ui-help-styles')) {
-    const style = document.createElement('style');
-    style.id = 'ui-help-styles';
-    style.textContent = `
-      .ui-help-container {
-        position: relative;
-        display: inline-block;
-        margin: 5px;
-      }
-      
-      .ui-help-button {
-        width: 24px;
-        height: 24px;
-        border-radius: 50%;
-        background-color: #6c757d;
-        color: white;
-        border: none;
-        font-size: 14px;
-        font-weight: bold;
-        cursor: pointer;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        transition: background-color 0.2s ease;
-      }
-      
-      .ui-help-button:hover {
-        background-color: #5a6268;
-      }
-      
-      .ui-help-tooltip {
-        position: fixed;
-        background-color: #333;
-        color: white;
-        padding: 8px 12px;
-        border-radius: 4px;
-        font-size: 16px;
-        white-space: pre-wrap;
-        max-width: 300px;
-        z-index: 10001;
-        visibility: hidden;
-        opacity: 0;
-        transition: opacity 0.2s ease;
-        box-shadow: 0 2px 8px rgba(0,0,0,0.2);
-        pointer-events: none;
-      }
-      
-
-      
-      .ui-help-popup {
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        background-color: rgba(0, 0, 0, 0.8);
-        z-index: 10000;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: 20px;
-        box-sizing: border-box;
-      }
-      
-      .ui-help-popup-content {
-        background-color: #2d3748;
-        color: #e2e8f0;
-        border-radius: 8px;
-        max-width: 90%;
-        max-height: 90%;
-        overflow-y: auto;
-        padding: 24px;
-        position: relative;
-        box-shadow: 0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04);
-        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-        line-height: 1.6;
-      }
-      
-      .ui-help-popup-close {
-        position: absolute;
-        top: 16px;
-        right: 16px;
-        background: none;
-        border: none;
-        color: #a0aec0;
-        font-size: 24px;
-        cursor: pointer;
-        padding: 4px;
-        line-height: 1;
-      }
-      
-      .ui-help-popup-close:hover {
-        color: #e2e8f0;
-      }
-      
-      .ui-help-popup-content h1,
-      .ui-help-popup-content h2,
-      .ui-help-popup-content h3 {
-        color: #f7fafc;
-        margin-top: 0;
-        margin-bottom: 16px;
-      }
-      
-      .ui-help-popup-content h1 { font-size: 1.875rem; }
-      .ui-help-popup-content h2 { font-size: 1.5rem; }
-      .ui-help-popup-content h3 { font-size: 1.25rem; }
-      
-      .ui-help-popup-content p {
-        margin: 16px 0;
-        color: #cbd5e0;
-      }
-      
-      .ui-help-popup-content ul,
-      .ui-help-popup-content ol {
-        padding-left: 24px;
-        margin: 16px 0;
-      }
-      
-      .ui-help-popup-content li {
-        margin: 8px 0;
-        color: #cbd5e0;
-      }
-      
-      .ui-help-popup-content code {
-        background-color: #4a5568;
-        color: #f7fafc;
-        padding: 2px 6px;
-        border-radius: 4px;
-        font-family: 'JetBrains Mono', 'Fira Code', 'Courier New', monospace;
-        font-size: 0.875rem;
-      }
-      
-      .ui-help-popup-content pre {
-        background-color: #1a202c;
-        color: #f7fafc;
-        padding: 16px;
-        border-radius: 6px;
-        overflow-x: auto;
-        margin: 16px 0;
-        border: 1px solid #4a5568;
-      }
-      
-      .ui-help-popup-content pre code {
-        background-color: transparent;
-        padding: 0;
-        color: inherit;
-      }
-      
-      .ui-help-popup-content a {
-        color: #63b3ed;
-        text-decoration: none;
-      }
-      
-      .ui-help-popup-content a:hover {
-        color: #90cdf4;
-        text-decoration: underline;
-      }
-      
-      .ui-help-popup-content strong {
-        color: #f7fafc;
-        font-weight: 600;
-      }
-      
-      .ui-help-popup-content em {
-        color: #e2e8f0;
-        font-style: italic;
-      }
-      
-      .ui-help-popup-content blockquote {
-        border-left: 4px solid #4a5568;
-        margin: 16px 0;
-        padding-left: 16px;
-        color: #a0aec0;
-        font-style: italic;
-      }
-    `;
-    document.head.appendChild(style);
-  }
-
-  return helpContainer;
-}
-
-function showHelpPopup(htmlContent) {
-  // Remove any existing popup
-  const existingPopup = document.querySelector('.ui-help-popup');
-  if (existingPopup) {
-    existingPopup.remove();
-  }
-
-  // Create popup
-  const popup = document.createElement('div');
-  popup.className = 'ui-help-popup';
-
-  const popupContent = document.createElement('div');
-  popupContent.className = 'ui-help-popup-content';
-
-  const closeButton = document.createElement('button');
-  closeButton.className = 'ui-help-popup-close';
-  closeButton.innerHTML = '&times;';
-  closeButton.setAttribute('aria-label', 'Close help');
-
-  const contentDiv = document.createElement('div');
-  contentDiv.innerHTML = htmlContent;
-
-  popupContent.appendChild(closeButton);
-  popupContent.appendChild(contentDiv);
-  popup.appendChild(popupContent);
-
-  // Add event listeners
-  closeButton.addEventListener('click', () => {
-    popup.remove();
-  });
-
-  popup.addEventListener('click', (e) => {
-    if (e.target === popup) {
-      popup.remove();
-    }
-  });
-
-  // Handle escape key
-  const handleEscape = (e) => {
-    if (e.key === 'Escape') {
-      popup.remove();
-      document.removeEventListener('keydown', handleEscape);
-    }
-  };
-  document.addEventListener('keydown', handleEscape);
-
-  // Add to DOM
-  document.body.appendChild(popup);
-}
-
-function showTooltip(button, tooltip) {
-  // Get button position relative to viewport
-  const buttonRect = button.getBoundingClientRect();
-
-  // Make tooltip visible but transparent to measure it
-  tooltip.style.visibility = 'visible';
-  tooltip.style.opacity = '0';
-
-  // Now get tooltip dimensions
-  const tooltipRect = tooltip.getBoundingClientRect();
-
-  // Calculate tooltip position
-  const buttonCenterX = buttonRect.left + buttonRect.width / 2;
-  const tooltipTop = buttonRect.top - tooltipRect.height - 8; // 8px gap above button
-
-  // Position tooltip centered above the button
-  let tooltipLeft = buttonCenterX - tooltipRect.width / 2;
-
-  // Ensure tooltip doesn't go off-screen horizontally
-  const padding = 10;
-  if (tooltipLeft < padding) {
-    tooltipLeft = padding;
-  } else if (tooltipLeft + tooltipRect.width > window.innerWidth - padding) {
-    tooltipLeft = window.innerWidth - tooltipRect.width - padding;
-  }
-
-  // Position tooltip
-  tooltip.style.left = `${tooltipLeft}px`;
-  tooltip.style.top = `${tooltipTop}px`;
-
-  // Show tooltip with fade-in
-  tooltip.style.opacity = '1';
-}
-
-function hideTooltip(tooltip) {
-  tooltip.style.visibility = 'hidden';
-  tooltip.style.opacity = '0';
 }
 
 /**
@@ -2390,46 +1672,7 @@ export class JsonUiManager {
   }
 }
 
-// Global debug and configuration controls for UI Manager
-if (typeof window !== 'undefined') {
-  window.setUiDebug = function (enabled = true) {
-    // Access the global UI manager instance if available
-    if (window.uiManager && typeof window.uiManager.setDebugMode === 'function') {
-      window.uiManager.setDebugMode(enabled);
-    } else {
-      console.warn(
-        '🎵 UI Manager instance not found. Debug mode will be applied when manager is created.',
-      );
-      // Store the preference for when the manager is created
-      window._pendingUiDebugMode = enabled;
-    }
-  };
-
-  window.setUiSpilloverThresholds = function (config) {
-    if (window.uiManager && typeof window.uiManager.updateSpilloverConfig === 'function') {
-      window.uiManager.updateSpilloverConfig(config);
-    } else {
-      console.warn(
-        '🎵 UI Manager instance not found. Spillover config will be applied when manager is created.',
-      );
-      window._pendingSpilloverConfig = config;
-    }
-  };
-
-  window.getUiSpilloverThresholds = function () {
-    if (window.uiManager && typeof window.uiManager.getSpilloverConfig === 'function') {
-      return window.uiManager.getSpilloverConfig();
-    }
-    console.warn('🎵 UI Manager instance not found.');
-    return null;
-  };
-
-  // Example usage helper
-  window.setUiSpilloverExample = function () {
-    console.log('🎵 Example spillover configuration:');
-    console.log('setUiSpilloverThresholds({');
-    console.log('  twoContainer: { minGroups: 3, minElements: 6, minElementsPerGroup: 2 },');
-    console.log('  threeContainer: { minGroups: 5, minElements: 10, minElementsPerGroup: 2 }');
-    console.log('});');
-  };
-}
+// Install the global debug and configuration controls for UI Manager. The
+// implementation lives in `./ui_debug_api.ts`; calling it here preserves the
+// original module-load behavior of `ui_manager.ts`.
+installUiDebugApi();
